Refactor SyncActivity to extract provisioning logic
Extract the directory-setup/download-kickoff block into a single
provisionApp function in ConfigurationsRepository, so SyncActivity
only makes a single call.

Fixes #15587
